% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_s3vec}
\alias{is_s3vec}
\title{Test for an s3vec object}
\usage{
is_s3vec(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
In order for an object to be considered a proper s3vec object, it has to
}
\details{
\itemize{
\item all child elements have to be of the same class
\item the "s3vec" class attribute has to be in last position
\item the remaining class attributes have to be the ones of the child
elements
}
}
\examples{
a <- structure("a", class = "foo")
b <- structure("b", class = "foo")

is_s3vec(s3vec(a, b))  # TRUE
is.s3vec(s3vec(a, b))  # TRUE

is_s3vec(a)            # FALSE

}
