% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3vec.R
\name{s3vec}
\alias{s3vec}
\alias{new_s3vec}
\title{Construct an s3vec object}
\usage{
s3vec(...)

new_s3vec(x)
}
\arguments{
\item{...}{Individual s3 objects}

\item{x}{A list of s3 objects}
}
\value{
An s3vec object that wraps around the supplied s3 objects.
}
\description{
\code{new_s3vec} is a low-level constructor that takes a list of
s3 objects of the same class. \code{s3vec} constructs an s3vec from
individual s3 objects.
}
\examples{
a <- structure("a", class = "foo")
b <- structure("b", class = "foo")

new_s3vec(list(a, b))
s3vec(a, b)

}
